+++ CHANGELOG.md
@@ -33,6 +33,7 @@
   `QueryRequest<C: CustomQuery>`. This allows us to provide
   `impl<C: CustomQuery> From<C> for QueryRequest<C>` for any custom query.
 - Implement `From<Binary> for Vec<u8>`.
+- Implement `From<CanonicalAddr> for Vec<u8>`.
 
 **cosmwasm-storage**